Update container dependency precedence

+++ ci_cd/linux/build.sh
@@ -11,10 +11,10 @@ PKG_FILE_RPM="${PROJECTDIR}/artifacts/entrusted-linux-amd64-${APPVERSION}.rpm"
 
 mkdir -p ${ARTIFACTSDIR}
 
-rm -rf ${PROJECTDIR}/entrusted_container/target
-rm -rf ${PROJECTDIR}/entrusted_client/target
-rm -rf ${PROJECTDIR}/entrusted_webclient/target
-rm -rf ${PROJECTDIR}/entrusted_webserver/target
+test -d ${PROJECTDIR}/entrusted_container/target && rm -rf ${PROJECTDIR}/entrusted_container/target
+test -d ${PROJECTDIR}/entrusted_client/target && rm -rf ${PROJECTDIR}/entrusted_client/target
+test -d ${PROJECTDIR}/entrusted_webclient/target && rm -rf ${PROJECTDIR}/entrusted_webclient/target
+test -d ${PROJECTDIR}/entrusted_webserver/target && rm -rf ${PROJECTDIR}/entrusted_webserver/target
 
 cd ${PROJECTDIR}
 
@@ -22,12 +22,10 @@ echo "Building entrusted_client (entrusted-gui)"
 cp -f ${PROJECTDIR}/images/Entrusted.png ${SCRIPTDIR}/appdir/entrusted-gui.png
 
 podman run --rm --privileged -v "${PROJECTDIR}":/src -v "${SCRIPTDIR}/appdir":/appdir -v "${PROJECTDIR}/artifacts":/artifacts docker.io/uycyjnzgntrn/rust-centos7:1.60.0 /bin/bash -c "ln -sf /usr/lib64/libfuse.so.2.9.2 /usr/lib/libfuse.so.2 && mkdir -p /tmp/appdir/usr/bin /tmp/appdir/usr/share/icons && cp /src/ci_cd/linux/xdg/* /tmp/appdir/ && cd /src/entrusted_client && /root/.cargo/bin/cargo build --release --bin entrusted-gui && cp target/release/entrusted-gui /tmp/appdir/ && cp /src/images/Entrusted.png /tmp/appdir/usr/share/icons/entrusted-gui.png && ARCH=x86_64 linuxdeploy --appdir /tmp/appdir --desktop-file /tmp/appdir/entrusted-gui.desktop --icon-filename /tmp/appdir/usr/share/icons/entrusted-gui.png --output appimage && mv *.AppImage /artifacts/entrusted-linux-amd64-${APPVERSION}/Entrusted_GUI-x86_64.AppImage"
-echo "Restoring old GUI Desktop file to discard appimagetool changes"
-#cd ${PROJECTDIR} && git checkout ci_cd/linux/appdir/entrusted-gui.desktop && cd -
 
 retVal=$?
 if [ $retVal -ne 0 ]; then
-	echo "Failure"
+	echo "Failure to create Linux GUI AppImage binary"
   exit 1
 fi
 
@@ -37,7 +35,7 @@ podman run --rm --volume "${PWD}":/root/src --workdir /root/src docker.io/joselu
 
 retVal=$?
 if [ $retVal -ne 0 ]; then
-	echo "Failure"
+	echo "Fail to build other Linux CLI binaries"
   exit 1
 fi
